Extract plan_sections.rs from turn.rs (parent #680)
Move the plan-section progress helpers out of turn.rs into a new
src/agent/loop_run/plan_sections.rs sibling module:

- join_sections_for_progress (pub(super))
- plan_sections_with_content (pub(super))
- plan_section_body_for_progress (pub(super))
- plan_section_has_content (private)
- normalize_plan_heading_for_progress (private)

All public items are pub(super); no facade re-export (DR3-001). Callers
updated:
- actor_loop_flow.rs: switches `join_sections_for_progress` /
  `plan_section_body_for_progress` / `plan_sections_with_content` imports
  from `super::turn` to `super::plan_sections`.
- tool_display.rs: switches `join_sections_for_progress` import from
  `super::turn` to `super::plan_sections`.
- turn.rs: drops the now-unused re-imports (these fns are not called from
  turn.rs after the migration).

turn.rs: 27,018 -> 26,919 (-99 LOC)
Cumulative from 39,489: -12,570 LOC (-31.8%)

Verification:
- cargo fmt --check
- cargo clippy --lib --all-targets -- -D warnings
- cargo test --lib (3,114 passed)
