+++ src/agent/loop_run/actor_loop_flow.rs
@@ -57,6 +57,9 @@ use super::active_job_arbiter::{LoopControlAction, RecoveryDispatchGate, Recover
 use super::interrupt::{InterruptFlag, InterruptMonitor};
 use super::lifecycle;
 use super::path_helpers::normalize_exploration_path;
+use super::plan_sections::{
+    join_sections_for_progress, plan_section_body_for_progress, plan_sections_with_content,
+};
 use super::progress_text::{
     format_progress_field, paint, progress_available_width, sanitize_for_progress, tool_color,
     tool_emoji, truncate,
@@ -71,8 +74,7 @@ use super::tool_history::focused_edit_target_already_read;
 use super::tool_history::is_plan_file_tool_call;
 use super::tool_policy::EffectiveToolPolicy;
 use super::turn::{
-    LOG_ARGS_MAX_CHARS, PLAN_REPEATED_EXPLORATION_BLOCK_THRESHOLD, join_sections_for_progress,
-    plan_section_body_for_progress, plan_sections_with_content, write_stdout_rendered,
+    LOG_ARGS_MAX_CHARS, PLAN_REPEATED_EXPLORATION_BLOCK_THRESHOLD, write_stdout_rendered,
 };
 use crate::agent::prompting;
 use crate::session::compact::approximate_token_count;
